requirements.txt
playwright==1.*
requests>=2.31
google-api-python-client
google-auth
google-auth-httplib2